init: start battle ship task
