+++ index.ts
@@ -1,6 +1,9 @@
-import { httpServer } from "./src/http_server/index.js";
+import { httpServer } from './src/http_server/index';
+import { wsServer } from './src/ws_server/index';
 
 const HTTP_PORT = 8181;
 
 console.log(`Start static http server on the ${HTTP_PORT} port!`);
 httpServer.listen(HTTP_PORT);
+
+wsServer;
